"Failing" test case for regex denial-of-service via URL authority
This test case will make the tests appear to hang.  If let to run long
enough, the test should actually pass.  I didn't see a way using
elm-test to set a timeout on a test and ensure that the test case
finishes within that timeout.

In extractPath, the URL authority (aka host as called by Erl) is used
directly in the regex, without escaping, to match against itself for
removal.  The host part is untrusted user-input and shouldn't be used as
a regular expression itself.  A fix is in the next commit.